rename entry files to entrypoint
